New object handling, resolvables

+++ lib/src/client.dart
@@ -17,7 +17,8 @@ class Client {
     "messageDelete": [],
     "messageEdit": [],
     "debug": [],
-    "loginError": []
+    "loginError": [],
+    "guildCreate": []
   };
 
   /// The token passed into the constructor.
@@ -31,6 +32,15 @@ class Client {
 
   /// The bot's OAuth2 app.
   ClientOAuth2Application app;
+
+  /// A map of all the guilds the bot is in, by id.
+  Map<String, Guild> guilds = {};
+
+  /// A map of all the channels the bot is in, by id.
+  Map<String, Channel> channels = {};
+
+  /// Whether or not the client is ready.
+  bool ready = false;
 
   void _heartbeat() {
     this._socket.add(JSON.encode({"op": 1,"d": this._lastS}));
@@ -81,6 +91,10 @@ class Client {
         this._sessionID = json['d']['session_id'];
         this.user = new ClientUser(json['d']['user']);
 
+        json['d']['guilds'].forEach((o) {
+          this.guilds[o['id']] = null;
+        });
+
         if (this.user.bot) {
           this._api.headers['Authorization'] = "Bot ${this.token}";
 
@@ -93,31 +107,127 @@ class Client {
           this._api.headers['Authorization'] = this.token;
         }
 
-        this._handlers['ready'].forEach((function) => function());
+        //this._handlers['ready'].forEach((function) => function());
       }
 
       else if (json['t'] == "MESSAGE_CREATE") {
-        Message message = new Message(json['d']);
-        this._handlers['message'].forEach((function) => function(message));
-      }
-
-      else if (json['t'] == "MESSAGE_DELETE") {
-        this._handlers['messageDelete'].forEach((function) => function(json['d']['id'], json['d']['channel_id']));
-      }
-
-      else if (json['t'] == "MESSAGE_UPDATE") {
-        if (!json['d'].containsKey('embeds')) {
-          Message message = new Message(json['d']);
-          this._handlers['messageEdit'].forEach((function) => function(message));
+        if (this.ready) {
+          Message message = new Message(this, json['d']);
+          this._handlers['message'].forEach((function) => function(message));
         }
       }
 
-      /*else if (json['t'] == "GUILD_CREATE") {
-        Guild guild = new Guild(json['d'], true);
-        this.guilds[guild.id] =  guild;
+      else if (json['t'] == "MESSAGE_DELETE") {
+        if (this.ready) {
+          this._handlers['messageDelete'].forEach((function) => function(json['d']['id'], json['d']['channel_id']));
+        }
+      }
 
-        //this._handlers['message'].forEach((function) => function(message));
-      }*/
+      else if (json['t'] == "MESSAGE_UPDATE") {
+        if (this.ready) {
+          if (!json['d'].containsKey('embeds')) {
+            Message message = new Message(this, json['d']);
+            this._handlers['messageEdit'].forEach((function) => function(message));
+          }
+        }
+      }
+
+      else if (json['t'] == "GUILD_CREATE") {
+        Guild guild = new Guild(this, json['d'], true);
+        this.guilds[guild.id] = guild;
+
+        guild.channels.forEach((i, v) {
+          this.channels[v.id] = v;
+        });
+
+        if (!this.ready) {
+          bool match = true;
+          this.guilds.forEach((i, o) {
+            if (o == null) {
+              match = false;
+            }
+          });
+
+          if (match == true) {
+            this.ready = true;
+            this._handlers['ready'].forEach((function) => function());
+          }
+        } else {
+          this._handlers['guildCreate'].forEach((function) => function(guild));
+        }
+      }
+    }
+  }
+
+  String _resolve(String to, Object object) {
+    if (to == "channel") {
+      if (object is Message) {
+        return object.channel.id;
+      } else if (object is Channel) {
+        return object.id;
+      } else if (object is Guild) {
+        return object.defaultChannel.id;
+      } else {
+        return object;
+      }
+    }
+
+    else if (to == "message") {
+      if (object is Message) {
+        return object.id;
+      } else {
+        return object;
+      }
+    }
+
+    else if (to == "guild") {
+      if (object is Message) {
+        return object.guild.id;
+      } else if (object is Channel) {
+        return object.guild.id;
+      } else if (object is Guild) {
+        return object.id;
+      } else {
+        return object;
+      }
+    }
+
+    else if (to == "user") {
+      if (object is Message) {
+        return object.author.id;
+      } else if (object is User) {
+        return object.id;
+      } else if (object is Member) {
+        return object.user.id;
+      } else {
+        return object;
+      }
+    }
+
+    else if (to == "member") {
+      if (object is Message) {
+        return object.author.id;
+      } else if (object is User) {
+        return object.id;
+      } else if (object is Member) {
+        return object.user.id;
+      } else {
+        return object;
+      }
+    }
+
+    else if (to == "app") {
+      if (object is User) {
+        return object.id;
+      } else if (object is Member) {
+        return object.user.id;
+      } else {
+        return object;
+      }
+    }
+
+    else {
+      return null;
     }
   }
 
@@ -155,22 +265,28 @@ class Client {
   ///
   /// Throws an [Exception] if the HTTP request errored.
   ///     Client.sendMessage("channel id", "My content!");
-  Future<Message> sendMessage(String channel, String content, [MessageOptions options]) async {
-    if (options == null) {
-      options = new MessageOptions();
-    }
+  Future<Message> sendMessage(Object channel, String content, [MessageOptions options]) async {
+    if (this.ready) {
+      if (options == null) {
+        options = new MessageOptions();
+      }
 
-    if (options.disableEveryone || (options.disableEveryone == null && this.options.disableEveryone)) {
-      content = content.replaceAll("@everyone", "@\u200Beveryone").replaceAll("@here", "@\u200Bhere");
-    }
+      if (options.disableEveryone || (options.disableEveryone == null && this.options.disableEveryone)) {
+        content = content.replaceAll("@everyone", "@\u200Beveryone").replaceAll("@here", "@\u200Bhere");
+      }
 
-    var r = await this._api.post('channels/$channel/messages', {"content": content, "tts": options.tts, "nonce": options.nonce});
-    Map res = JSON.decode(r.body);
+      String id = this._resolve("channel", channel);
 
-    if (r.statusCode == 200) {
-      return new Message(res);
+      var r = await this._api.post('channels/$id/messages', {"content": content, "tts": options.tts, "nonce": options.nonce});
+      Map res = JSON.decode(r.body);
+
+      if (r.statusCode == 200) {
+        return new Message(this, res);
+      } else {
+        throw new Exception("${res['code']}: ${res['message']}");
+      }
     } else {
-      throw new Exception("${res['code']}: ${res['message']}");
+      throw new Exception("the client isn't ready");
     }
   }
 
@@ -178,13 +294,20 @@ class Client {
   ///
   /// Throws an [Exception] if the HTTP request errored.
   ///     Client.sendMessage("channel id", "message id");
-  Future<bool> deleteMessage(String channel, String message) async {
-    var r = await this._api.delete('channels/$channel/messages/$message');
+  Future<bool> deleteMessage(Object channel, Object message) async {
+    if (this.ready) {
+      channel = this._resolve('channel', channel);
+      message = this._resolve('message', message);
 
-    if (r.statusCode == 204) {
-      return true;
+      var r = await this._api.delete('channels/$channel/messages/$message');
+
+      if (r.statusCode == 204) {
+        return true;
+      } else {
+        throw new Exception("'deleteMessage' error.");
+      }
     } else {
-      throw new Exception("'deleteMessage' error.");
+      throw new Exception("the client isn't ready");
     }
   }
 
@@ -192,56 +315,58 @@ class Client {
   ///
   /// Throws an [Exception] if the HTTP request errored.
   ///     Client.sendMessage("channel id", "message id", "My edited content!");
-  Future<Message> editMessage(String channel, String message, String content) async {
-    var r = await this._api.patch('channels/$channel/messages/$message', {"content": content});
-    Map res = JSON.decode(r.body);
+  Future<Message> editMessage(Object channel, Object message, String content, [MessageOptions options]) async {
+    if (this.ready) {
+      if (options == null) {
+        options = new MessageOptions();
+      }
 
-    if (r.statusCode == 200) {
-      return new Message(res);
+      if (options.disableEveryone || (options.disableEveryone == null && this.options.disableEveryone)) {
+        content = content.replaceAll("@everyone", "@\u200Beveryone").replaceAll("@here", "@\u200Bhere");
+      }
+
+      channel = this._resolve('channel', channel);
+      message = this._resolve('message', message);
+
+      var r = await this._api.patch('channels/$channel/messages/$message', {"content": content});
+      Map res = JSON.decode(r.body);
+
+      if (r.statusCode == 200) {
+        return new Message(this, res);
+      } else {
+        throw new Exception("${res['code']}:${res['message']}");
+      }
     } else {
-      throw new Exception("${res['code']}:${res['message']}");
+      throw new Exception("the client isn't ready");
     }
   }
 
-  /// Gets a [Channel] object.
-  ///
-  /// Throws an [Exception] if the HTTP request errored.
-  ///     Client.getChannel("channel id");
-  Future<Channel> getChannel(String id) async {
-    var r = await this._api.get('channels/$id');
-    Map res = JSON.decode(r.body);
-    if (r.statusCode == 200) {
-      return new Channel(res);
-    } else {
-      throw new Exception("${res['code']}:${res['message']}");
-    }
-  }
-
-  /// Gets a [Guild] object.
-  ///
-  /// Throws an [Exception] if the HTTP request errored.
-  ///     Client.getGuild("guild id");
-  Future<Guild> getGuild(String id) async {
-    var r = await this._api.get('guilds/$id');
-    Map res = JSON.decode(r.body);
-    if (r.statusCode == 200) {
-      return new Guild(res, false);
-    } else {
-      throw new Exception("${res['code']}:${res['message']}");
-    }
-  }
-
-  /// Gets a [Member] object.
+  /// Gets a [Member] object. Adds it to `Client.guilds["guild id"].members` if
+  /// not already there.
   ///
   /// Throws an [Exception] if the HTTP request errored.
   ///     Client.getMember("guild id", "user id");
-  Future<Member> getMember(String guild, String member) async {
-    var r = await this._api.get('guilds/$guild/members/$member');
-    Map res = JSON.decode(r.body);
-    if (r.statusCode == 200) {
-      return new Member(res);
+  Future<Member> getMember(Object guild, Object member) async {
+    if (this.ready) {
+
+      guild = this._resolve('guild', guild);
+      member = this._resolve('member', member);
+
+      if (this.guilds[guild].members[member] != null) {
+        return this.guilds[guild].members[member];
+      } else {
+        var r = await this._api.get('guilds/$guild/members/$member');
+        Map res = JSON.decode(r.body);
+        if (r.statusCode == 200) {
+          Member m = new Member(res);
+          this.guilds[guild].members[m.user.id] = m;
+          return m;
+        } else {
+          throw new Exception("${res['code']}:${res['message']}");
+        }
+      }
     } else {
-      throw new Exception("${res['code']}:${res['message']}");
+      throw new Exception("the client isn't ready");
     }
   }
 
@@ -249,13 +374,19 @@ class Client {
   ///
   /// Throws an [Exception] if the HTTP request errored.
   ///     Client.getUser("user id");
-  Future<User> getUser(String id) async {
-    var r = await this._api.get('users/$id');
-    Map res = JSON.decode(r.body);
-    if (r.statusCode == 200) {
-      return new User(res);
+  Future<User> getUser(Object user) async {
+    if (this.ready) {
+      user = this._resolve('user', user);
+
+      var r = await this._api.get('users/$id');
+      Map res = JSON.decode(r.body);
+      if (r.statusCode == 200) {
+        return new User(res);
+      } else {
+        throw new Exception("${res['code']}:${res['message']}");
+      }
     } else {
-      throw new Exception("${res['code']}:${res['message']}");
+      throw new Exception("the client isn't ready");
     }
   }
 
@@ -264,12 +395,16 @@ class Client {
   /// Throws an [Exception] if the HTTP request errored.
   ///     Client.getInvite("invite code");
   Future<Invite> getInvite(String code) async {
-    var r = await this._api.get('invites/$code');
-    Map res = JSON.decode(r.body);
-    if (r.statusCode == 200) {
-      return new Invite(res);
+    if (this.ready) {
+      var r = await this._api.get('invites/$code');
+      Map res = JSON.decode(r.body);
+      if (r.statusCode == 200) {
+        return new Invite(res);
+      } else {
+        throw new Exception("${res['code']}:${res['message']}");
+      }
     } else {
-      throw new Exception("${res['code']}:${res['message']}");
+      throw new Exception("the client isn't ready");
     }
   }
 
@@ -278,17 +413,24 @@ class Client {
   /// Throws an [Exception] if the HTTP request errored or if the client user
   /// is not a bot.
   ///     Client.getMessage("channel id", "message id");
-  Future<Message> getMessage(String channel, String message) async {
-    if (this.user.bot) {
-      var r = await this._api.get('channels/$channel/messages/$message');
-      Map res = JSON.decode(r.body);
-      if (r.statusCode == 200) {
-        return new Message(res);
+  Future<Message> getMessage(Object channel, String message) async {
+    if (this.ready) {
+      if (this.user.bot) {
+        channel = this._resolve('channel', channel);
+        message = this._resolve('message', message);
+
+        var r = await this._api.get('channels/$channel/messages/$message');
+        Map res = JSON.decode(r.body);
+        if (r.statusCode == 200) {
+          return new Message(this, res);
+        } else {
+          throw new Exception("${res['code']}:${res['message']}");
+        }
       } else {
-        throw new Exception("${res['code']}:${res['message']}");
+        throw new Exception("'getMessage' is only usable by bot accounts.");
       }
     } else {
-      throw new Exception("'getMessage' is only usable by bot accounts.");
+      throw new Exception("the client isn't ready");
     }
   }
 
@@ -297,17 +439,23 @@ class Client {
   /// Throws an [Exception] if the HTTP request errored or if the client user
   /// is a bot.
   ///     Client.getOAuth2Info("app id");
-  Future<OAuth2Info> getOAuth2Info(String id) async {
-    if (!this.user.bot) {
-      var r = await this._api.get('oauth2/authorize?client_id=$id&scope=bot');
-      Map res = JSON.decode(r.body);
-      if (r.statusCode == 200) {
-        return new OAuth2Info(res);
+  Future<OAuth2Info> getOAuth2Info(Object app) async {
+    if (this.ready) {
+      if (!this.user.bot) {
+        app = this._resolve('app', app);
+
+        var r = await this._api.get('oauth2/authorize?client_id=$app&scope=bot');
+        Map res = JSON.decode(r.body);
+        if (r.statusCode == 200) {
+          return new OAuth2Info(res);
+        } else {
+          throw new Exception("${res['code']}:${res['message']}");
+        }
       } else {
-        throw new Exception("${res['code']}:${res['message']}");
+        throw new Exception("'getOAuth2Info' is only usable by user accounts.");
       }
     } else {
-      throw new Exception("'getOAuth2Info' is only usable by user accounts.");
+      throw new Exception("the client isn't ready");
     }
   }
 
@@ -316,17 +464,24 @@ class Client {
   /// Throws an [Exception] if the HTTP request errored or if the client user
   /// is a bot.
   ///     Client.oauth2Authorize("app id", "guild id");
-  Future<bool> oauth2Authorize(String app, String guild, [int permissions = 0]) async {
-    if (!this.user.bot) {
-      var r = await this._api.post('oauth2/authorize?client_id=$app&scope=bot', {"guild_id": guild, "permissions": permissions, "authorize": true});
-      Map res = JSON.decode(r.body);
-      if (r.statusCode == 200) {
-        return true;
+  Future<bool> oauth2Authorize(Object app, Object guild, [int permissions = 0]) async {
+    if (this.ready) {
+      if (!this.user.bot) {
+        app = this._resolve('app', app);
+        guild = this._resolve('guild', guild);
+
+        var r = await this._api.post('oauth2/authorize?client_id=$app&scope=bot', {"guild_id": guild, "permissions": permissions, "authorize": true});
+        Map res = JSON.decode(r.body);
+        if (r.statusCode == 200) {
+          return true;
+        } else {
+          throw new Exception("${res['code']}:${res['message']}");
+        }
       } else {
-        throw new Exception("${res['code']}:${res['message']}");
+        throw new Exception("'oauth2Authorize' is only usable by user accounts.");
       }
     } else {
-      throw new Exception("'oauth2Authorize' is only usable by user accounts.");
+      throw new Exception("the client isn't ready");
     }
   }
 }
